Make dinamically generated 16x16 grid using flex
"main" tag was added in index.html to contain all
divs, which are created after via javascript.

Each square (div) is equally sized on the screen.
Their sizes are calculated dinamically on script.js
so that a 16x16 grid is shown.

The color chosen (purple) serves just to view what
is going on visually.