Add a button to get the user's input
A button has been added in the index.html and
an event was attached to it in script.js.
Everytime it is pressed, a prompt shows asking
the user to input a new dimension for the grid.

For debugging purposes, nothing special happens.
If the input is not valid, it is printed this
info on the console. Otherwise, it is printed that
the number passed is valid.