Create the grid based on user's input
An interesting feature was added to the button.
Everytime it's clicked, it asks the user a
new dimension for the grid, which can't be
less than 1 and more than 100. Once the input
is given, the existing tiles are removed and
new ones are created.

Instead of setting the tiles' properties in
the script.js file, a class is attached to them,
so that the script code turns itself more
readable. The class' name is "tile".

It is no longer necessary to calculate each
tile's dimension. This responsibility is now
passed to the flex algorithm. To make this
possible, <main> tag has the "display: flex"
and "flex-direction: column" properties. Besides,
each tile is contained inside an outside div, which
would simulate the lines of the grid. These
outside divs have a class named ".tiles_container"
attached to them.